moved player to separate module to unclutter main
