+++ src/player.hpp
@@ -1,0 +1,19 @@
+/**
+ * @file player.hpp
+ * @author melektron
+ * @brief music player for the create
+ * @version 0.1
+ * @date 2023-03-10
+ * 
+ * @copyright Copyright FrenchBakery (c) 2023
+ * 
+ */
+
+#pragma once
+
+
+namespace Player
+{
+    void start();
+    void stop();
+}
